amend the code on the new demo app
